backend/requirements.txt
# Kitchen Assistant Application Requirements
# Python 3.12+ recommended

# Web Server
Flask==3.0.0
Werkzeug==3.0.1

# HTTP Requests
requests==2.31.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==2.1.0

# Environment Variables
python-dotenv==1.0.0

# JSON Handling
simplejson==3.19.2

# Logging and Debugging
rich==13.7.0

